+++ src/System.Diagnostics.FileVersionInfo/tests/System.Diagnostics.FileVersionInfo.Tests/FileVersionInfoTest.cs
@@ -2,136 +2,17 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Diagnostics;
-using System.Globalization;
 using System.IO;
-using System.Runtime.InteropServices;
 using System.Text;
 using Xunit;
 
 namespace System.Diagnostics.Tests
 {
-    public class FileVersionInfoTest
+    public partial class FileVersionInfoTest : FileCleanupTestBase
     {
-        private const string NativeConsoleAppFileName = "NativeConsoleApp.exe";
-        private const string NativeLibraryFileName = "NativeLibrary.dll";
-        private const string SecondNativeLibraryFileName = "SecondNativeLibrary.dll";
         private const string TestAssemblyFileName = "System.Diagnostics.FileVersionInfo.TestAssembly.dll";
         private const string TestCsFileName = "Assembly1.cs";
         private const string TestNotFoundFileName = "notfound.dll";
-
-        [Fact]
-        [PlatformSpecific(TestPlatforms.Windows)] // native PE files only supported on Windows
-        public void FileVersionInfo_Normal()
-        {
-            // NativeConsoleApp (English)
-            VerifyVersionInfo(Path.Combine(Directory.GetCurrentDirectory(), NativeConsoleAppFileName), new MyFVI()
-            {
-                Comments = "",
-                CompanyName = "Microsoft Corporation",
-                FileBuildPart = 3,
-                FileDescription = "This is the description for the native console application.",
-                FileMajorPart = 5,
-                FileMinorPart = 4,
-                FileName = Path.Combine(Directory.GetCurrentDirectory(), NativeConsoleAppFileName),
-                FilePrivatePart = 2,
-                FileVersion = "5.4.3.2",
-                InternalName = NativeConsoleAppFileName,
-                IsDebug = false,
-                IsPatched = false,
-                IsPrivateBuild = false,
-                IsPreRelease = true,
-                IsSpecialBuild = true,
-                Language = GetFileVersionLanguage(0x0409), //English (United States)
-                LegalCopyright = "Copyright (C) 2050",
-                LegalTrademarks = "",
-                OriginalFilename = NativeConsoleAppFileName,
-                PrivateBuild = "",
-                ProductBuildPart = 3,
-                ProductMajorPart = 5,
-                ProductMinorPart = 4,
-                ProductName = Path.GetFileNameWithoutExtension(NativeConsoleAppFileName),
-                ProductPrivatePart = 2,
-                ProductVersion = "5.4.3.2",
-                SpecialBuild = ""
-            });
-        }
-
-        [Fact]
-        [PlatformSpecific(TestPlatforms.Windows)] // native PE files only supported on Windows
-        public void FileVersionInfo_Chinese()
-        {
-            // NativeLibrary.dll (Chinese)
-            VerifyVersionInfo(Path.Combine(Directory.GetCurrentDirectory(), NativeLibraryFileName), new MyFVI()
-            {
-                Comments = "",
-                CompanyName = "A non-existent company",
-                FileBuildPart = 3,
-                FileDescription = "Here is the description of the native library.",
-                FileMajorPart = 9,
-                FileMinorPart = 9,
-                FileName = Path.Combine(Directory.GetCurrentDirectory(), NativeLibraryFileName),
-                FilePrivatePart = 3,
-                FileVersion = "9.9.3.3",
-                InternalName = "NativeLibrary.dll",
-                IsDebug = false,
-                IsPatched = true,
-                IsPrivateBuild = false,
-                IsPreRelease = true,
-                IsSpecialBuild = false,
-                Language = GetFileVersionLanguage(0x0004),//Chinese (Simplified)
-                Language2 = GetFileVersionLanguage(0x0804),//Chinese (Simplified, PRC) - changed, but not yet on all platforms
-                LegalCopyright = "None",
-                LegalTrademarks = "",
-                OriginalFilename = "NativeLibrary.dll",
-                PrivateBuild = "",
-                ProductBuildPart = 40,
-                ProductMajorPart = 20,
-                ProductMinorPart = 30,
-                ProductName = "I was never given a name.",
-                ProductPrivatePart = 50,
-                ProductVersion = "20.30.40.50",
-                SpecialBuild = "",
-            });
-        }
-
-        [Fact]
-        [PlatformSpecific(TestPlatforms.Windows)] // native PE files only supported on Windows
-        public void FileVersionInfo_DifferentFileVersionAndProductVersion()
-        {
-            // Mtxex.dll
-            VerifyVersionInfo(Path.Combine(Directory.GetCurrentDirectory(), SecondNativeLibraryFileName), new MyFVI()
-            {
-                Comments = "",
-                CompanyName = "",
-                FileBuildPart = 0,
-                FileDescription = "",
-                FileMajorPart = 0,
-                FileMinorPart = 65535,
-                FileName = Path.Combine(Directory.GetCurrentDirectory(), SecondNativeLibraryFileName),
-                FilePrivatePart = 2,
-                FileVersion = "0.65535.0.2",
-                InternalName = "SecondNativeLibrary.dll",
-                IsDebug = false,
-                IsPatched = false,
-                IsPrivateBuild = false,
-                IsPreRelease = false,
-                IsSpecialBuild = false,
-                Language = GetFileVersionLanguage(0x0400),//Process Default Language
-                LegalCopyright = "Copyright (C) 1 - 2014",
-                LegalTrademarks = "",
-                OriginalFilename = "SecondNativeLibrary.dll",
-                PrivateBuild = "",
-                ProductBuildPart = 0,
-                ProductMajorPart = 1,
-                ProductMinorPart = 0,
-                ProductName = "Unknown_Product_Name",
-                ProductPrivatePart = 1,
-                ProductVersion = "1.0.0.1",
-                SpecialBuild = "",
-            });
-        }
 
         [Fact]
         public void FileVersionInfo_CustomManagedAssembly()
@@ -154,7 +35,7 @@ namespace System.Diagnostics.Tests
                 IsPrivateBuild = false,
                 IsPreRelease = false,
                 IsSpecialBuild = false,
-                Language = RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? GetFileVersionLanguage(0x0000) : "Language Neutral",
+                Language = GetFileVersionLanguage(0x0000),
                 LegalCopyright = "Copyright, you betcha!",
                 LegalTrademarks = "TM",
                 OriginalFilename = TestAssemblyFileName,
@@ -206,7 +87,14 @@ namespace System.Diagnostics.Tests
         }
 
         [Fact]
-        public void FileVersionInfo_FileNotFound()
+        public void FileVersionInfo_CurrentDirectory_FileNotFound()
+        {
+            Assert.Throws<FileNotFoundException>(() =>
+            FileVersionInfo.GetVersionInfo(Directory.GetCurrentDirectory()));
+        }
+
+        [Fact]
+        public void FileVersionInfo_NonExistentFile_FileNotFound()
         {
             Assert.Throws<FileNotFoundException>(() =>
                 FileVersionInfo.GetVersionInfo(Path.Combine(Directory.GetCurrentDirectory(), TestNotFoundFileName)));
@@ -217,7 +105,7 @@ namespace System.Diagnostics.Tests
         // [] DLL has unknown codepage info
         // [] DLL language/codepage is 8-hex-digits (locale > 0x999) (different codepath)
 
-        private void VerifyVersionInfo(String filePath, MyFVI expected)
+        private void VerifyVersionInfo(string filePath, MyFVI expected)
         {
             FileVersionInfo fvi = FileVersionInfo.GetVersionInfo(filePath);
             Assert.Equal(expected.Comments, fvi.Comments);
@@ -249,7 +137,7 @@ namespace System.Diagnostics.Tests
             Assert.Equal(expected.SpecialBuild, fvi.SpecialBuild);
 
             //ToString
-            String nl = Environment.NewLine;
+            string nl = Environment.NewLine;
             Assert.Equal("File:             " + fvi.FileName + nl +
                       "InternalName:     " + fvi.InternalName + nl +
                       "OriginalFilename: " + fvi.OriginalFilename + nl +
@@ -340,13 +228,6 @@ namespace System.Diagnostics.Tests
             }
             buffer.Append("\"");
             return (buffer.ToString());
-        }
-
-        private static string GetFileVersionLanguage(uint langid)
-        {
-            var lang = new StringBuilder(256);
-            Interop.Kernel32.VerLanguageName(langid, lang, (uint)lang.Capacity);
-            return lang.ToString();
         }
     }
 }
